Cleanup Command single quotes redundant gsub
Remove a block of code which was in place prior to cb1a3c83, where
where powershell nested quote escape substitution moved to another
part of the script composition. The original code block was hit
subsequently forcing a double escape of the already doubled single
quotes (turning a "'" input into "''''" by the end of it).