Rework quote handling
Testing shows that the way single and double quotes are handled in
psexec context across different platforms/versions of PSH appears
to differ. It may be an encoding issue in our pipeline, or actually
a difference on the target host. Recent changes addressing quote
handling help on Win10/16 platforms, but break things in earlier
versions without specific options being set at the Msf namespace
level.

Rewire the handling of quotes to always offer inner-wraping on the
inner command invocation with single quotes and escaped single
quotes as appropriate, while permitting the double quote wrapping
of the outer command context as is apparently required by newer
versions of windows due to spaces (powershell -c "command string").

Testing:
  Internal lab tests against Server 2016 and 2008r2.

Notes:
  Psexec on Win2016 has significant delays. Likely due to the
attempted introspection of execution by defender. Consistent
acquisition of shells was accomplished with handlers separate from
the exploit module context with the handler disabled in the module
itself, or by using WfsDelay of 90+