Improve use of RandomIdentifier::Generator
Following along with logic in rex-rig #3, implement options for
RIGs used by PSH-related code to exclude reserved variables, drop
variable sizes down from a max of 12 to 5, and permit passing an
existing RIG to the Script initializer.

Implement the options as a constant in Script and Templates.
Include Templates in Payload, use the included constants to provide
the default template_path parameters to paylaod generators.